+++ scripts/changelog.ts
@@ -304,8 +304,9 @@ async function exec(): Promise<void> {
         '**plugin-rows**. add-button is now always visible under the focused document'
       ],
       added: [
-        '**plugin-serlo-injection**. Added injection plugin for serlo.org content',
-        '**plugin-table**. You can now create tables in markdown.'
+        '**core**. Add `StateType.migratable` that allows to update plugin state structure without breaking changes',
+        '**plugin-serlo-injection**. Add plugin for injecting serlo.org content',
+        '**plugin-table**. Add plugin for markdown tables'
       ],
       fixed: ['**plugin-text**. Split plugin when pasting multiple blocks']
     }